requirements.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile --max-rounds=30 requirements.in
#
asn1crypto==1.5.1
    # via
    #   oscrypto
    #   snowflake-connector-python
boto3==1.11.10
    # via -r requirements.in
botocore==1.14.10
    # via
    #   boto3
    #   s3transfer
certifi==2021.10.8
    # via
    #   requests
    #   snowflake-connector-python
cffi==1.15.0
    # via
    #   cryptography
    #   snowflake-connector-python
charset-normalizer==2.0.12
    # via
    #   requests
    #   snowflake-connector-python
click==7.0
    # via -r requirements.in
cryptography==36.0.2
    # via
    #   pyopenssl
    #   snowflake-connector-python
docutils==0.15.2
    # via botocore
idna==3.3
    # via
    #   requests
    #   snowflake-connector-python
jmespath==0.9.4
    # via
    #   boto3
    #   botocore
oscrypto==1.3.0
    # via snowflake-connector-python
pycparser==2.21
    # via cffi
pycryptodomex==3.14.1
    # via snowflake-connector-python
pyjwt==2.3.0
    # via snowflake-connector-python
pyopenssl==21.0.0
    # via snowflake-connector-python
python-dateutil==2.8.1
    # via botocore
pytz==2022.1
    # via snowflake-connector-python
requests==2.27.1
    # via snowflake-connector-python
s3transfer==0.3.2
    # via boto3
six==1.14.0
    # via
    #   pyopenssl
    #   python-dateutil
snowflake-connector-python==2.7.7
    # via -r requirements.in
urllib3==1.25.8
    # via
    #   botocore
    #   requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
